Extract client method for devices and device events in DeviceClient class
